refactor(core): extract load_triplets and split greedy decode steps
Move parallel session triplet loading and encoder_model_path into
load_files. Split greedy_decode_impl into select_token and
commit_non_blank helpers without changing decode behavior (existing
unit tests cover bias, blank-run, and endpoint paths).
